Fixed all of the incorrect uses of \item as: \item [item-name]; they're now all \item[item-name], which makes 'em all look consistent.
New macro: \JikesTM{}.  Similar ones for SPEC, Java, Linux, and other
words that are registered or unregistered trademarks.  They obsolete
the \*TMFootnote macros.

The \*TMFooter{} macros are now no-ops; they're still in the text in
case we want to go back to the old way of handling trademark notes.

"Red Hat" is two words, not one.  It's also a trademark.  IBM is a
trademark.  Intel is a trademark.  As are SPEC, SPECjbb, SPECjvm,
Unix, and SuSE.  All of these corrected and marked as such.

Added more index entries to the FAQ.

We now consistently refer to "GNU Classpath" when we're speaking of
The GNU Classpath libraries specifically.

More documentation on the building glitches people have had with GNU
Classpath.

More documentation on how and why we use the preprocessor.