All changes to the Bash version-checking code.
Expanded and clarified the error message that users see when they use
Bash 2.05, so that it reflect the fact that the problems with Bash
2.05 can disappear until a trivial change in the code makes them reappear.

Fixed a bug that would have cropped up if BASH_VERSINFO were ever
unset.  Of course, it's never supposed to be unset, but this is a
sanity check, so it makes sense to guard against the condition.

Fixed an incomplete comment.

Re-ordered functions to put the entry point to this code at the top of
the section.