Fixed a write buffer exhaustion bug.
a)  The code did not properly check for a zero return value
    on write buffer slow path.
b)  There was no heuristic to avoid inordinately long mutator
    phases which would generate massive write buffers

Fixed a) with a suitable check, fixed b) with a heuristic which
triggers a GC once more than half teh available virtual memory
is consumed (a lot of remset entries!).

Because this error was not being handled nicely, and because I
never run things with such huge heaps and unbounded nursery, a) I've
not encountered this before, and b) it fails in an ugly way when
assertions are off.  So, it is possible that this is one of our
long standing GC bugs...   Here's hoping.