This patch changes the way that the architecture dependent code accesses some of the common facilities. Prior to this patch the architecture independent and architecture dependent code both used ArchitectureSpecific.Foo where as after this patch the architecture dependent code will now use <arch>.Foo where possible and architecture independent code will use ArchitectureSpecific.Foo.
This has the advantage that less code will be red in an IDE if you are editing ppc code where the ArchitectureSpecific has been generated for ia32 and vice versa. This is the first step towards getting the ppc and ia32 code being able to coexist and be compiled together in one tree.