Parallel hardware traps were not working correctly on linux/ia32. When using processors 2 (or higher) and a program that causes hardware traps on multiple threads close together in time, we could abruptly fail with a segfault.
What was happening was we were calling pthread_mutex_lock from the signal handler. We can not call this method from here! When the lock is busy, this
method *blocks* which means our thread can be given another signal to handle. I am not 100% sure of the details here, I am not a signal handling
expert, but I have implemented a solution for now where we spin on pthread_mutex_trylock. This causes the problems to disappear. I have added a unit
test (only meaningful on BaseBase builds) to core/threads that checks for this problem.

Someone should look at if this lock is necessary at all though.

This fixes the persistent lusearch bug, so now we should be able to run all of dacapo with multiple virtual processors!