Fix for baseline GC maps. After we have refined the reference map as at the start of a try block due to a branch, we must propagate this information into all of the handlers and re-iterate through them.
If we do not, we risk lowering a REF in a handler block to a NONREF.
This can of course cause us to process an invalid map.

This corrects the longstanding (years old?) JBB GC map bug. Getting this
right in all the cases involving JSRs was tricky. I used the full
commit test run to try and ensure I didn't miss anything.

This should make it in time for the sanity runs starting shortly, I will
investigate any new regressions. It would be valuable for anyone else
that knows this code well to double check that what I am doing makes
sense.

We are really close to having zero regressions on SPECjvm98,
SPECjbb2000, SPECjbb2005 and dacapo on Intel with SSE2.