Added an initial version of buildit (unfortunately written in perl).
This is a script that we have used at ANU for a while cleaned up and
contributed back. The basic idea is to make it easier to build and test
the RVM on multiple machines.

Be careful with it though, as it is doing some scary rsync --delete
operations. Please make use of the --dry-run until you are happy with
what the script is doing. With --dry-run it echos the commands it is
running.

I guessed to put this in bin, but it wasn't a principled decision.

Suggestions and comments welcome!