Bug fix; sleepInternal was incorrectly ignoring the ns argument when computing its wakeup time.
To investigate:  without this change, I was getting reliable hangs on jbb2000 on a development image with 4 virtual processors on an 8-way blade.  Dumps of the VM state showed that the problem was that all the jbb worker threads were in sleep with wakupNanoTime values that were clearly bogus (snippet appended).  I can see no reason (other than an opt compiler bug) that adding in the ns parameter should have changed this, but it appears that it might have.  I was getting the hang 100% of the time before, I have now run it 6 times in a row without the hang with this one change in place.

-- System Queues --
 wakeupQueue: nowNano=4630161740834 (0x000004360b255422)
1-SLEEPING(wakeupNano=291627975102269632.00)20-SLEEPING(wakeupNano=291627975102271552.00)21-SLEEPING(wakeupNano=291627975102686144.00)19-SLEEPING(wake
upNano=291627975106104448.00)
 debuggerQueue:
 collectorQueue: 7-daemon-collector-BLOCKED 8-daemon-collector-BLOCKED 6-daemon-collector-BLOCKED 9-daemon-collector-BLOCKED
 finalizerQueue: 10-daemon-BLOCKED