RVM-191.  Reduce estimate of codespace cost of getstatic of a final primitive field because we expect that by the time the opt compiler ever compiles such a method, the field will be initialized and we will be able to make the contents of the field a simple constant operand instead of actually generating code to load the value.
To support this, an additional stage was added to bootimage writing.  After all types in the bootimage are resolved, we give ourselves a chance to do additional analyses before we actually compile any code (so far, this stage is only doing the trivial re-run of the methodSummary computation of all bootimage methods).  We should look at the various interprocedural summaries we are computing before compilation and see if there is anything else that should be moved to this stage of bootimage writing.