Back out of 13776 and 13777.
There seems to be a significant GC performance problem introduced by 13776.  By backing out we can determine whether subsequent commits have any performance issues in isolation.  This also allows other commits to roll forward.  Obviously it is easy enough for anyone to reverse this decision if they feel it appropriate.  For the moment though, it seems to me to be the sensible thing to do.