RVM-357.  Fix baseline compilers so that the generate an explicit null check before calling the uninterruptible locking code for the monitorenter bytecode.  Change necessary because we've somewhat recently made raising a NullPointerException in uninterruptible code a fatal error.
Note:  We don't need to do an explicit nullcheck when we're calling the locking code in a method prologue, because the object cannot possibly be null there (otherwise we wouldn't have reached the prologue).  Ditto for epilogue.

Technically, we should probably do a nullcheck on monitorexit too, but that can only happen when the bytecodes are not from Java (ie, when the monitorexit is happening on a lock that hasn't been entered). Going to ignore that fringe case until it is actually seen in the wild...