RVM-352.
The code sequence used by the opt compiler to implement fp remainder was unbalanced in its use of the 387 floating point stack (two pushes, one pop).  Eventually this results in an fp stackoverflow exception (unless the FP state happens to get cleared for some other reason).  Add an ffree into the codesequence for FPREM so that the 387 stack is balanced.

Kudos to Ian for doing all the detective work to find the problem.