small cleanup of Java==>native method transition sequence: eliminate nativeMethodInvoker stub from OutOfLineMachineCode.
When the GC could relocate code, we needed a non-moving stub to actually make the call into the native code because we couldn't update the return address in the native stack if GC relocated the code.  Now that code spaces are non-moving, we don't need this extra stub call.  This change "inlines" and slightly simplifies the instruction sequence that was in the stub into the native methods generated by the JNI compiler. I think we could further optimize the sequence now that it's all in one place, but will look at that in a separate commit.