RVM-746 WIP : Clone and minimally modify org.jikesrvm.classloader.Primitive to create org.jikesrvm.classloader.UnboxedType.
This commit is fairly trivial and would not be useful in and of itself, since it doesn't actually change how we recognize UnboxedTypes and compile their magic methods.  The goal in committing it is to make sure that there aren't any pieces of code outside of the classloader that still think Address et al are instances of Primitives (instead of UnboxedType).

The next step is to actually change UnboxedType to be created from a java classfile that has the @Unboxed annotation and to build the method & fields of UnboxedType based on the information in the class file.