Minor comment fixes in Gen and PretenureThresholdFraction.
JavaDoc interprets something like

/***********************
  * Constants
  */

which is used in Gen as a JavaDoc comment. This results in the constant SURVIVAL_ESTIMATE in Gen getting the JavaDoc comment "Constants" which isn't what we want. Therefore, insert empty /* */ lines in Gen so that SURVIVAL_ESTIMATE doesn't get any JavaDoc. Conversely, no such empty comment is needed in PretenureThresholdFraction because there isn't any JavaDoc comment that could be erroneously copied.