RVM-1041 : Add a unit test for org.jikesrvm.compilers.opt.bc2ir.GenerationContext.
The unit test was written in a way that avoids changes to GenerationContext. As a consequence, the test methods are not always independent (especially those checking the GenerationContext constructors).

The test case does not test functionality that relates to edge counts because the edge count code is not clearly separated (see RVM-1052).