Fix more problems in checkstylePlugin.xml (thanks to Dave).
The ant property verifyproperty of the checksum task must be set to the name of the property and not its contents (see https://ant.apache.org/manual/Tasks/checksum.html).

By default, the fail task checks only if a property is set (or not set). We need to check the contents of the property because we always set it.