Fix JavaDoc errors in the IA32 Assembler.
Most of the errors were caused by the lack of escaping for characters like &, < and >. The rest of the errors were caused by incorrect usage of HTML tags.

Note that the function for escaping in genAssembler.sh is only used for the emitBinary* methods. The emitFloat* methods do not use any operators that need escaping.