Only show output for eager mmapping of discontigous spaces when the verbosity level is high enough.
The old behaviour is inconsistent with the eager mmaping of contigous spaces which only shows printouts when the verbosity level is > 2.

Moreover, the option -X:gc:eagerMmapSpaces=true may be necessary on some platforms to ensure that the operating system and MMTk have the same view about available memory. On those platforms, it is useful if there's no additional output when using -X:gc:eagerMmapSpaces=true because that means that test cases with checksum-based output verification work without adjustments.