RVM-1066 : Change bootimage runner to correctly recognize -X:bootclasspath/a and -X:bootclasspath/p as non-standard command line arguments.
The -Xbootclasspath/a and -Xbootclasspath/p options were introduced in commit 72e6d186bcc31d598f643a2ead84acc79693ecb4 to fix RVM-621. Unfortunately, the original commit did not add them to nonStandardArgs in cmdLine.h. This caused the bootimage runner to treat them as application arguments instead of non-standard arguments.

All non-standard arguments are expected to appear before the first application argument. The appearance of any application argument communicates to the bootimage runner that it can safely hand the rest of the arguments to the VM and that does not need to continue to check for non-standard arguments. This in turn caused an option order like "rvm -Xbootclasspath/a:doesnotexist -Xmx1G" to fail with "vm: "-Xmx1G" is not a recognized Jikes RVM command line argument"  instead of the expected  "vm: Please specify a class to execute.  vm:   You can invoke the VM with the "-help" flag for usage information.".