RVM-1080 : Detect invalid operations on unboxed types in the optimizing compiler.
It would be preferable to detect invalid operations on unboxed types in a compiler-independent way, e.g. in a bytecode verifier. We currently don't have a suitable component for this. Therefore, I've decided to detect the invalid operations in BC2IR which already contains code for the detection of illegal upcasts of unboxed types.