RVM-1079 : Mark the constant interface GCSpyPlan as deprecated.
IMHO GCSpyPlan does not really seem to fit the current design of MMTk where variables for spaces are kept in the respective plan classes. On the other hand, I don't think that a conversion of GCSpyPlan to a constant class would improve the design. Therefore, I've elected to mark the class as deprecated which will make sure that it jumps out as problem if anybody works on updating GCSpy.