Use explicit iteration in SpillLocationManager to be able to properly remove elements during iteration.
Note that both the OpenJDK and GNU Classpath seem to throw ConcurrentModificationException only if next() is called on the iterator after a removal from the backing data structure. This does not happen in the changed code fragment because the loop is terminated immediately after the removal. This behaviour seems to be intended in GNU Classpath (see https://gcc.gnu.org/bugzilla/show_bug.cgi?id=27028 ).