RVM-1083 : Save machine code offsets in a separate class instead of using Instruction's scratch field and remove the now unused field.
This commit also introduces changes to cope with two instances of missing machine code offsets that have been revealed by the move to a separate data structure for the offset.

Firstly, machine code offsets are not set for IR_PROLOGUE instructions in non-interruptible code. This is merely an oddity and it is not easily fixable, so the code works around that by dealing with IR_PROLOGUE instructions in a special way.

Secondly, when the exception table is generated in OptExceptionTable, the code may encounter catch blocks that have no associated code (and therefore no machine code offsets) because they are unreachable according to the optimizing compiler. This occurs when building a development boot image in the method writeSFManifest(..) from gnu.java.util.jar.JarUtils . In such a case, it is safe to not add the entry to the table because it is impossible that the entry would be used. This is based on the assumption that the optimizing compiler was correct in determining that the catch block is unreachable. In any case, this behaviour is no worse than the previous behaviour of falsely claiming that there is a catch block at machine code offset 0, which is directly at the start of the method.

The rest of the changes are pretty straightforward with the exception of some changes to the assemblers. It was necessary to save the machine code offsets in the assemblers via an instance variable which in turn made it necessary to move away from having a static generateCode() method as entrypoint for code generation.