Fix uninterruptibility problems in the new code for checking for unreachable catch blocks that are reached at runtime.
Note that the pre-commit run did not detect these problems because protopye doesn't include the code (because it's in the opt compiler) and development does not check for uninterruptibility problems (because the checks are only implemented in the baseline compilers). I neglected to build prototype-opt manually which could have detected these problems.