Add support for the JUnit test format in regression tests.
The "JUnit test format" is the XML format that's used by the Ant JUnit task. AFAIK there is no formal specification of the format so we use what is understood by Jenkins. The alternatives would have been to write a Jenkins plugin that understands our format or to use another general format (e.g. TAP which is also supported via a Jenkins plugin, see https://wiki.jenkins-ci.org/display/JENKINS/TAP+Plugin ). Both of these options were likely to require more work and it is unclear to me if they would have offered substantial advantages for the increased effort.

It's not possible to do a 1-to-1 mapping of our format to the JUnit format. For example, the JUnit format provides a possibility to distinguish between the output of System.out and System.err . Such a distinction does not exist in our test results. Conversely, our test format contains some additional information (e.g. command lines, exit codes, parameters) that cannot be mapped onto the JUnit format. This problem is currently solved by putting the additional information from our reports into the JUnit output for System.err and the names of the testsuites.