Backed out multiple changesets related to RVM-1083 because they belong to a line of development that isn't salvageable.
The changeset af1e1b490a33822e0cf6d45bf714be856a786861 (RVM-1083 : Save guard information in GenerationContext.) was the final changeset in that line of development.

The commits
- 10986:fa4df3929879119ab88523b7d41eb6cc94a6c2bd (RVM-1083 : Move static methods that deal with guards from BC2IR to GenerationContext.)
- 10987:ddbf1fe8c73bac8f7ef045b6ebf27cafdf722cc7 (RVM-1083 : Convert methods that deal with guards in GenerationContext to instance methods.)
- 10988:6a62e43b1efa4caa7c57b3c8b1a87144dba1baed (RVM-1083 : Change all accesses to guards in the HIR to go through GenerationContext.)
are part of changes to move the responsibility for saving guards from scratchObject in RegisterOperand to GenerationContext.

The changes were incorrect because they did not properly deal with the guard information in RegisterOperand. The guard information was not propagated to copies of RegisterOperand made via copy(). The resulting lack of guard information caused more inlining which lead to an increase in the size of the bootimages.

It is impractical to change all the places where copy() is used just to pull the guard information out of RegisterOperand. Therefore, this line of development needed to be abandoned. It is much saner to just leave the guard information in RegisterOperand.