RVM-876 : Avoid compare-and-swap operations when objects are being forwarded (original changes by Laurence Hellyer).
In case the object is in the state of BEING_FORWARDED, it is unnecessary to try to change the bits to BEING_FORWARDED. If the object's state is FORWARDED, the state must not be changed to BEING_FORWARDED. Therefore, the compare-and-swap operation is only necessary when the object was not forwarded yet and it is not in the process of being forwarded, i.e. when forwarding has not been triggered yet.