Run static initializer for gnu.java.nio.charset.EncodingHelper earlier to fix an exception that wasn't visible without -X:verbose because the GNU Classpath code caught it.
VMClassLoader transitively uses EncodingHelper during the call to getBootPackages(). The call reads a manifest for a JAR file using an InputStreamReader with an UTF-8 encoding. The InputStreamReader constructor InputStreamReader(InputStream, String) determines the encoding from the String and tries to canonicalize the encoding. The attempt to find a canonical encoding caused an exception in EncodingHelper because that requires the canonical list which is built by the static initializer of EncodingHelper.