Improve documentation for JNIGenericHelpers.strlen(..) and provide an alternative implementation that avoids reads of undefined memory.
The strlen(..) method reads whole words by default. These words are then examined to find the terminating 0 byte for the C string. A part of the read word may be undefined which will cause warnings like "Conditional jump or move depends on uninitialised value(s)" in tools like Valgrind. However, the implementation of the method is safe and its behaviour is well-defined, so no problems occur.